+++ templates/fullstack-nextjs/eslint.config.js
@@ -1,0 +1,14 @@
+const js = require("@eslint/js");
+
+module.exports = [
+  js.configs.recommended,
+  {
+    rules: {
+      "no-unused-vars": ["warn", { argsIgnorePattern: "^_" }],
+      "no-console": "warn",
+    },
+  },
+  {
+    ignores: [".next/", "node_modules/"],
+  },
+];